influ-web/influ-backend/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "cloudinary>=1.44.1",
    "fastapi>=0.128.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.7",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.0",
    "pinecone>=7.3.0",
    "psycopg>=3.3.2",
    "psycopg-pool>=3.3.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "requests>=2.32.5",
    "sendgrid>=6.12.5",
    "trafilatura>=2.0.0",
    "uvicorn>=0.40.0",
]
